Context.set(newGlobalValue)
Adds a method to update a context's default (global) value. This schedules
an update on all context consumers that are not nested inside a provider,
across all roots and renderers. For most use cases, this replaces the need
to inject a context provider at the top of each root.

(I've added an `unstable_` prefix until it's ready for public release.)